Add files via upload
Dear editor
I am pleased to present our original manuscript Web system: Quick Response (QR) Code for Teacher Attendance Control at Educational Institution No. 20712 Santa Sana. Tana – 2024 at the University of the Province of Cañete, for consideration for publication as a research article. We believe that this manuscript is suitable for publication since the research focus reflects the objectives and scope of its journal. This manuscript presents the implementation of a web system to improve efficiency in the attendance of teaching staff, with the purpose of improving the accuracy of records and the facilitation of improving teaching attention and optimizing attendance data management. Where the entry and exit time, the number of weekly, monthly and annual hours that each teacher must teach can be reliably recorded. Meanwhile, the system created will allow efficient work to be carried out, creating a friendly interface, with many options and tools, with pleasant designs according to the requirements of EIE No. 2712 “Santa Ana de Tana”, Yauyos Province, Lima department , Peru.